Ratio endpoint (#15)
* implement ratio endpoint

* add react dockerfiles

* add base React UI

+++ api/responses/base.go
@@ -3,7 +3,7 @@ package responses
 // BaseResponseOK represents base reponse of successful request
 type BaseResponseOK struct {
 	Success bool        `json:"success"`
-	Code    uint8       `json:"code"`
+	Code    int         `json:"code"`
 	Data    interface{} `json:"data"`
 }
 
@@ -11,14 +11,14 @@ type BaseResponseOK struct {
 type BaseResponseError struct {
 	Success bool   `json:"success"`
 	Message string `json:"message"`
-	Code    uint16 `json:"code"`
+	Code    int    `json:"code"`
 }
 
-func makeResponseOK(succ bool, code uint8, data interface{}) BaseResponseOK {
+func makeResponseOK(succ bool, code int, data interface{}) BaseResponseOK {
 	return BaseResponseOK{Success: succ, Code: code, Data: data}
 }
 
 // MakeResponseError creates a error json response
-func MakeResponseError(succ bool, msg string, code uint16) BaseResponseError {
+func MakeResponseError(succ bool, msg string, code int) BaseResponseError {
 	return BaseResponseError{Success: succ, Message: msg, Code: code}
 }
